refactor: delete question feature and tests

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import type { QuestionsRepository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
+	authorId: string
 	questionId: string
 }
 
@@ -10,12 +11,17 @@ export class DeleteQuestionUseCase {
 	constructor(private questionsRepository: QuestionsRepository) {}
 
 	async execute({
+		authorId,
 		questionId,
 	}: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
 		const question = await this.questionsRepository.findById(questionId)
 
 		if (!question) {
 			throw new Error('Question not found.')
+		}
+
+		if (authorId !== question.authorId.toString()) {
+			throw new Error('Not allowed.')
 		}
 
 		await this.questionsRepository.delete(question)
